Use absolute dates in API calls.
The tool was using stream-local dates in API calls, for convenience to
users mostly. For time zones with daylight saving time, however, the
handling of hours on DST changes was not correct. Now the tool uses
absolute (UTC) dates internally for all API calls, translating to local
dates for display as needed.