Allow a minimum time gap constraint for creating Reset records.
Change the option to generate reset records to
--generate-reset-datum-min-gap that accepts a duration value, treated as
a minimum length of time to consider creating Reset records for data
gaps. The reasoning here is that an aggregation could just be "stale"
still because SolarNetwork changed its default reading difference
tolderance from 3 months to 1 year at one point, so it is possible that
the cached aggregations just need to be reprocessed, without needing a
reset record.