Update reset record generation logic.
The reset record generation suffered from some performance issues, so
reworked in the following ways. Now when bisecting ranges, we always
narrow the half ranges to the available data. This fixes the issue where
a half range in a large data gap producing invalid assesments. This then
simplifies the reset record logic, as we already have rough start/end
dates of the data gap as the end/start of the left/right bisect ranges.